기술 통계 API 수정 (#78)
* bug fix

* 통계관련 API 수정

+++ src/main/java/com/trendyTracker/Domain/Jobs/Statistics/Dto/ChartInfo.java
@@ -11,5 +11,5 @@ public record ChartInfo(
     @Schema(description = "완료 날짜",example = "2023-12-01")
     LocalDate toDate,
     @Schema(description = "차트 데이터",example = "HashMap")
-    HashMap<String, Integer> chart
+    HashMap<String, Long> chart
 ){}
